gm fix and changelog

+++ scripts/hooks.js
@@ -3,6 +3,7 @@ import { createDependentRegion, getFullFlagPath } from "./helpers.js";
 
 export default function registerHooks() {
     Hooks.on('createMeasuredTemplate', async (templateDoc) => {
+        if (!game.user.isGM) return;
         let originItem = await fromUuid(templateDoc.getFlag('dnd5e', 'origin'));
         if (!originItem) return;
         if (!(originItem.getFlag(CONSTANTS.MODULE_NAME, CONSTANTS.FLAGS.ATTACH_REGION_TO_TEMPLATE) ?? false)) return;
@@ -10,6 +11,7 @@ export default function registerHooks() {
     });
 
     Hooks.on('updateMeasuredTemplate', async (templateDoc) => {
+        if (!game.user.isGM) return;
         let region = await fromUuid(templateDoc.getFlag(CONSTANTS.MODULE_NAME, CONSTANTS.FLAGS.ATTACHED_REGION));
         if (!region) return;
         let origShape = templateDoc.object.shape;
@@ -26,6 +28,7 @@ export default function registerHooks() {
     });
     
     Hooks.on('deleteMeasuredTemplate', async (templateDoc) => {
+        if (!game.user.isGM) return;
         let region = await fromUuid(templateDoc.getFlag(CONSTANTS.MODULE_NAME, CONSTANTS.FLAGS.ATTACHED_REGION));
         if (!region) return;
         await region.delete();
